+++ tests/metadata/test_insar_gamma.py
@@ -19,6 +19,7 @@ def test_create_insar_gamma_readme(insar_product_dir):
         processor_name='GAMMA',
         processor_version='20191203',
         ref_point_coords={'lat': 64.8378, 'lon': -147.7164},
+        phase_filter_parameter=0.4,
     )
     writer = insar.InSarMetadataWriter(payload)
     output_file = writer.create_readme()
@@ -41,6 +42,7 @@ def test_create_amp_xml(insar_product_dir):
         processor_name='GAMMA',
         processor_version='20191203',
         ref_point_coords={'lat': 64.8378, 'lon': -147.7164},
+        phase_filter_parameter=0.4,
     )
     writer = insar.InSarMetadataWriter(payload)
     output_file = writer.create_amp_xml()
@@ -63,6 +65,7 @@ def test_create_coherence_xml(insar_product_dir):
         processor_name='GAMMA',
         processor_version='20191203',
         ref_point_coords={'lat': 64.8378, 'lon': -147.7164},
+        phase_filter_parameter=0.4,
     )
     writer = insar.InSarMetadataWriter(payload)
     output_file = writer.create_coherence_xml()
@@ -85,6 +88,7 @@ def test_create_dem_xml(insar_product_dir):
         processor_name='GAMMA',
         processor_version='20191203',
         ref_point_coords={'lat': 64.8378, 'lon': -147.7164},
+        phase_filter_parameter=0.4,
     )
     writer = insar.InSarMetadataWriter(payload)
     output_file = writer.create_dem_tif_xml()
@@ -107,6 +111,7 @@ def test_create_los_displacement_xml(insar_product_dir):
         processor_name='GAMMA',
         processor_version='20191203',
         ref_point_coords={'lat': 64.8378, 'lon': -147.7164},
+        phase_filter_parameter=0.4,
     )
     writer = insar.InSarMetadataWriter(payload)
     output_file = writer.create_los_displacement_xml()
@@ -129,6 +134,7 @@ def test_create_look_vector_xmls(insar_product_dir):
         processor_name='GAMMA',
         processor_version='20191203',
         ref_point_coords={'lat': 64.8378, 'lon': -147.7164},
+        phase_filter_parameter=0.4,
     )
     writer = insar.InSarMetadataWriter(payload)
     output_files = writer.create_look_vector_xmls()
@@ -154,6 +160,7 @@ def test_create_browse_xmls(insar_product_dir):
         processor_name='GAMMA',
         processor_version='20191203',
         ref_point_coords={'lat': 64.8378, 'lon': -147.7164},
+        phase_filter_parameter=0.4,
     )
     writer = insar.InSarMetadataWriter(payload)
     output_files = writer.create_browse_xmls()
@@ -179,6 +186,7 @@ def test_unwrapped_phase_xml(insar_product_dir):
         processor_name='GAMMA',
         processor_version='20191203',
         ref_point_coords={'lat': 64.8378, 'lon': -147.7164},
+        phase_filter_parameter=0.4,
     )
     writer = insar.InSarMetadataWriter(payload)
     output_file = writer.create_unwrapped_phase_xml()
@@ -201,6 +209,7 @@ def test_vertical_displacement_xml(insar_product_dir):
         processor_name='GAMMA',
         processor_version='20191203',
         ref_point_coords={'lat': 64.8378, 'lon': -147.7164},
+        phase_filter_parameter=0.4,
     )
     writer = insar.InSarMetadataWriter(payload)
     output_file = writer.create_vertical_displacement_xml()
@@ -223,6 +232,7 @@ def test_wrapped_phase_xml(insar_product_dir):
         processor_name='GAMMA',
         processor_version='20191203',
         ref_point_coords={'lat': 64.8378, 'lon': -147.7164},
+        phase_filter_parameter=0.4,
     )
     writer = insar.InSarMetadataWriter(payload)
     output_file = writer.create_wrapped_phase_xml()
@@ -245,6 +255,7 @@ def test_inc_map_xml(insar_product_dir):
         processor_name='GAMMA',
         processor_version='20191203',
         ref_point_coords={'lat': 64.8378, 'lon': -147.7164},
+        phase_filter_parameter=0.4,
     )
     writer = insar.InSarMetadataWriter(payload)
     output_file = writer.create_inc_map_xml()
@@ -267,6 +278,7 @@ def test_inc_map_ell_xml(insar_product_dir):
         processor_name='GAMMA',
         processor_version='20191203',
         ref_point_coords={'lat': 64.8378, 'lon': -147.7164},
+        phase_filter_parameter=0.4,
     )
     writer = insar.InSarMetadataWriter(payload)
     output_file = writer.create_inc_map_ell_xml()
@@ -289,6 +301,7 @@ def test_water_mask_xml(insar_product_dir):
         processor_name='GAMMA',
         processor_version='20191203',
         ref_point_coords={'lat': 64.8378, 'lon': -147.7164},
+        phase_filter_parameter=0.4,
     )
     writer = insar.InSarMetadataWriter(payload)
     output_file = writer.create_water_mask_xml()
@@ -311,6 +324,7 @@ def test_insar_gamma_all_files(insar_product_dir):
         processor_name='GAMMA',
         processor_version='20191203',
         ref_point_coords={'lat': 64.8378, 'lon': -147.7164},
+        phase_filter_parameter=0.4,
     )
     assert files == [
         insar_product_dir / 'S1AB_20210424T125204_20210430T125122_HHP006_INT80_G_ueF_B4A1.README.md.txt',